Fix 64/32 bit's integer(switch on size_t) Windows build fix(dll)

+++ RTF++/inc/Implementation/RTD/Containers/SharedVector.h
@@ -20,7 +20,7 @@ namespace RTF{
             SharedVector(const SharedVector& vector):   data(vector.data),
                                             lengthOfVector(vector.lengthOfVector),
                                             allocatedMemory(vector.allocatedMemory) {}
-            SharedVector(const T* data, uint64_t length): data(data), lengthOfVector(length) {}
+            SharedVector(const T* data, size_t length): data(data), lengthOfVector(length) {}
             SharedVector(SharedVector&& vector) noexcept :  data(vector.data), lengthOfVector(vector.lengthOfVector),
                                                 allocatedMemory(vector.allocatedMemory), stepOfAllocate(vector.stepOfAllocate){
                 vector.data.DestructThis();
@@ -76,13 +76,13 @@ namespace RTF{
                 }
                 return *this;
             }
-            SharedVector& operator()(T* data, uint64_t length){
+            SharedVector& operator()(T* data, size_t length){
                 this->data = data;
                 this->lengthOfVector = length;
                 this->allocatedMemory = length;
                 return *this;
             }
-            SharedVector& operator()(const T* data, uint64_t length){
+            SharedVector& operator()(const T* data, size_t length){
                 this->data = data;
                 this->lengthOfVector = length;
                 return *this;
@@ -90,7 +90,7 @@ namespace RTF{
             bool Empty() const{
                 return this->lengthOfVector == 0;
             }
-            uint64_t Size() const{
+            size_t Size() const{
                 return this->lengthOfVector;
             }
 
@@ -101,11 +101,11 @@ namespace RTF{
                 this->operator[](this->lengthOfVector) = pushData;
             }
 
-            void SetSize(uint64_t newSize){
+            void SetSize(size_t newSize){
                 Types::ASharedPointer<T> temporaryVariable(new T[(newSize + 1) - ((newSize + 1) % this->stepOfAllocate) + this->stepOfAllocate]);
-                uint64_t min = this->lengthOfVector < newSize ? this->lengthOfVector : newSize;
+                size_t min = this->lengthOfVector < newSize ? this->lengthOfVector : newSize;
 
-                for(uint64_t i = 0; i < min; i++)
+                for(size_t i = 0; i < min; i++)
                     temporaryVariable[i] = this->data[i];
 
                 this->allocatedMemory = (newSize + 1) - ((newSize + 1) % this->stepOfAllocate) + this->stepOfAllocate;
@@ -119,14 +119,14 @@ namespace RTF{
 
                 Types::ASharedPointer<T> tmp(new T[this->lengthOfVector - (this->lengthOfVector % this->stepOfAllocate) + this->stepOfAllocate]);
 
-                for(uint64_t j = 0; j < this->lengthOfVector; j++)
+                for(size_t j = 0; j < this->lengthOfVector; j++)
                     tmp[j] = this->data[j];
 
                 this->data = tmp;
                 this->allocatedMemory = this->lengthOfVector - (this->lengthOfVector % this->stepOfAllocate) + this->stepOfAllocate;
             }
 
-            void Erase(uint64_t start, uint64_t end){
+            void Erase(size_t start, size_t end){
                 if(start == 0 && end == this->lengthOfVector){
                     this->data = new T[this->stepOfAllocate];
                     this->allocatedMemory = this->stepOfAllocate;
@@ -137,7 +137,7 @@ namespace RTF{
                     return;
                 }
                 if(start > end){
-                    uint64_t tmp = start;
+                    size_t tmp = start;
                     start = end;
                     end = tmp;
                 }
@@ -145,7 +145,7 @@ namespace RTF{
                 int newAllocation = (newSize + 1) - ((newSize + 1) % this->stepOfAllocate) + this->stepOfAllocate;
                 Types::ASharedPointer<T> temporaryVariable = new T[newAllocation];
 
-                uint64_t i = 0;
+                size_t i = 0;
                 for(; i < start; i++){
                     temporaryVariable[i] = this->data[i];
                 }
@@ -158,12 +158,12 @@ namespace RTF{
                 this->allocatedMemory = newAllocation;
             }
 
-            T& operator[](uint64_t i){
+            T& operator[](size_t i){
                 i++;
                 if(this->data.IsLiteral() || this->allocatedMemory < i || !this->data.Alone()){
-                    uint64_t size;
+                    size_t size;
                     {
-                        uint64_t tmp1, tmp2;
+                        size_t tmp1, tmp2;
                         tmp1 = i - (i % this->stepOfAllocate) + this->stepOfAllocate;
                         tmp2 = this->allocatedMemory;
                         size = tmp1 > tmp2? tmp1 : tmp2;
@@ -171,7 +171,7 @@ namespace RTF{
 
                     Types::ASharedPointer<T> tmp(new T[size]);
 
-                    for(uint64_t j = 0; j < this->lengthOfVector; j++)
+                    for(size_t j = 0; j < this->lengthOfVector; j++)
                         tmp[j] = this->data[j];
 
                     this->data = tmp;
@@ -182,7 +182,7 @@ namespace RTF{
                 return this->data[i - 1];
             }
 
-            T GetElement(uint64_t i) const{
+            T GetElement(size_t i) const{
                 return this->data.GetElement(i);
             }
 
@@ -193,8 +193,8 @@ namespace RTF{
             ~SharedVector() = default;
         private:
             Types::ASharedPointer<T> data;
-            uint64_t lengthOfVector = 0;
-            uint64_t allocatedMemory = 0;
+            size_t lengthOfVector = 0;
+            size_t allocatedMemory = 0;
             uint16_t stepOfAllocate = 25;
         };
     }
