+++ RTF++/inc/Implementation/RTD/Strings/SharedString.h
@@ -37,14 +37,14 @@ namespace RTF {
             bool operator!=(const SharedString& string);
             bool operator!=(const char* str);
 
-            char& operator[](uint64_t i);
+            char& operator[](size_t i);
 
             explicit operator const char*();
 
             SharedString& ToUpper();
             SharedString& ToLower();
 
-            uint64_t Size() const;
+            size_t Size() const;
 
             ~SharedString();
